falcon_backend/requirements-dev.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.2
filelock==3.15.4
identify==2.6.0
iniconfig==2.0.0
JSON-log-formatter==0.5.1
Markdown==3.6
marshmallow==3.21.3
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
pytest==8.2.2
python-dotenv==1.0.1
python-logstash==0.4.8
PyYAML==6.0.1
ruff==0.5.1
sqlparse==0.5.0
virtualenv==20.26.3
